extract useDeepLinks hook
