+++ src/lib/AndroidApp.tsx
@@ -1,27 +1,25 @@
 import { ActionSheetProvider } from "@expo/react-native-action-sheet"
 import { GlobalStore, GlobalStoreProvider } from "lib/store/GlobalStore"
 import { Theme } from "palette"
-import React, { useEffect, useRef } from "react"
-import { useCallback } from "react"
-import { Linking, UIManager, View } from "react-native"
+import React, { useEffect } from "react"
+import { UIManager, View } from "react-native"
+import RNBootSplash from "react-native-bootsplash"
 import track from "react-tracking"
 import { RelayEnvironmentProvider } from "relay-hooks"
 import { useWebViewCookies } from "./Components/ArtsyReactWebView"
 import { _FancyModalPageWrapper } from "./Components/FancyModal/FancyModalContext"
 import { ToastProvider } from "./Components/Toast/toastHook"
 import { useSentryConfig } from "./ErrorReporting"
+import { ArtsyNativeModule } from "./NativeModules/ArtsyNativeModule"
 import { ModalStack } from "./navigation/ModalStack"
-import { navigate } from "./navigation/navigate"
 import { defaultEnvironment } from "./relay/createEnvironment"
 import { BottomTabsNavigator } from "./Scenes/BottomTabs/BottomTabsNavigator"
 import { ForceUpdate } from "./Scenes/ForceUpdate/ForceUpdate"
 import { Onboarding } from "./Scenes/Onboarding/Onboarding"
 import { AdminMenuWrapper } from "./utils/AdminMenuWrapper"
+import { useDeepLinks } from "./utils/useDeepLinks"
 import { ProvideScreenDimensions } from "./utils/useScreenDimensions"
 import { useStripeConfig } from "./utils/useStripeConfig"
-
-import RNBootSplash from "react-native-bootsplash"
-import { ArtsyNativeModule } from "./NativeModules/ArtsyNativeModule"
 
 if (UIManager.setLayoutAnimationEnabledExperimental) {
   UIManager.setLayoutAnimationEnabledExperimental(true)
@@ -30,12 +28,12 @@ if (UIManager.setLayoutAnimationEnabledExperimental) {
 const Main: React.FC<{}> = track()(({}) => {
   const isHydrated = GlobalStore.useAppState((state) => state.sessionState.isHydrated)
   const isLoggedIn = GlobalStore.useAppState((state) => !!state.auth.userAccessToken)
-  const launchURL = useRef<string | null>(null)
   const forceUpdateMessage = GlobalStore.useAppState((state) => state.config.echo.forceUpdateMessage)
 
   useSentryConfig()
   useStripeConfig()
   useWebViewCookies()
+  useDeepLinks()
 
   useEffect(() => {
     if (isHydrated) {
@@ -46,45 +44,6 @@ const Main: React.FC<{}> = track()(({}) => {
       }, 500)
     }
   }, [isHydrated])
-
-  useEffect(() => {
-    Linking.getInitialURL().then((url) => {
-      if (url) {
-        handleDeepLink(url)
-      }
-    })
-  }, [])
-
-  useEffect(() => {
-    Linking.addEventListener("url", ({ url }) => handleDeepLink(url))
-    return () => {
-      Linking.removeAllListeners("url")
-    }
-  }, [])
-
-  const handleDeepLink = useCallback(
-    (url: string) => {
-      // If the state is hydrated and the user is logged in
-      // We navigate them to the the deep link
-      if (isHydrated && isLoggedIn) {
-        navigate(url)
-      }
-
-      // Otherwise, we save the deep link url
-      // to redirect them to the login screen once they log in
-      launchURL.current = url
-    },
-    [isHydrated, isLoggedIn]
-  )
-
-  useEffect(() => {
-    if (isLoggedIn && launchURL.current) {
-      // Navigate to the saved launch url
-      navigate(launchURL.current)
-      // Reset the launchURL
-      launchURL.current = null
-    }
-  }, [isLoggedIn, launchURL.current])
 
   if (!isHydrated) {
     return <View></View>
