+++ src/lib/utils/useDeepLinks.ts
@@ -1,0 +1,49 @@
+import { navigate } from "lib/navigation/navigate"
+import { GlobalStore } from "lib/store/GlobalStore"
+import { useCallback, useEffect, useRef } from "react"
+import { Linking } from "react-native"
+
+export function useDeepLinks() {
+  const isLoggedIn = GlobalStore.useAppState((state) => !!state.auth.userAccessToken)
+  const isHydrated = GlobalStore.useAppState((state) => state.sessionState.isHydrated)
+  const launchURL = useRef<string | null>(null)
+
+  useEffect(() => {
+    Linking.getInitialURL().then((url) => {
+      if (url) {
+        handleDeepLink(url)
+      }
+    })
+  }, [])
+
+  useEffect(() => {
+    Linking.addEventListener("url", ({ url }) => handleDeepLink(url))
+    return () => {
+      Linking.removeAllListeners("url")
+    }
+  }, [])
+
+  const handleDeepLink = useCallback(
+    (url: string) => {
+      // If the state is hydrated and the user is logged in
+      // We navigate them to the the deep link
+      if (isHydrated && isLoggedIn) {
+        navigate(url)
+      }
+
+      // Otherwise, we save the deep link url
+      // to redirect them to the login screen once they log in
+      launchURL.current = url
+    },
+    [isHydrated, isLoggedIn]
+  )
+
+  useEffect(() => {
+    if (isLoggedIn && launchURL.current) {
+      // Navigate to the saved launch url
+      navigate(launchURL.current)
+      // Reset the launchURL
+      launchURL.current = null
+    }
+  }, [isLoggedIn, launchURL.current])
+}
